+++ packages/jest-expo/CHANGELOG.md
@@ -11,6 +11,7 @@
 ### 💡 Others
 
 - Improve local mock lookup. ([#39743](https://github.com/expo/expo/pull/39743) by [@aleqsio](https://github.com/aleqsio))
+- Add mocks for pedometer Recording API functions and event update helpers. ([#35652](https://github.com/expo/expo/pull/35652) by [@Krisztiaan](https://github.com/Krisztiaan))
 - Remove `experiments.reactCanary` support in favor of built-in React +19.1 support. ([#40386](https://github.com/expo/expo/pull/40386) by [@EvanBacon](https://github.com/EvanBacon))
 - Fixed check-packages error on Windows. ([#41194](https://github.com/expo/expo/pull/41194) by [@kudo](https://github.com/kudo))
 - Switch `react-server-dom-webpack` to optional peer dependency ([#41379](https://github.com/expo/expo/pull/41379) by [@kitten](https://github.com/kitten))